+++ R/transform.R
@@ -177,12 +177,13 @@ as_hm <- function(x)
 #' as_integer(rnorm(10))
 #' as_integer(letters)
 #' as_integer(5:10 %>% as.factor)
-#' as.integer(5:10 %>% as.factor)
+#' as.integer(c("a","b") %>% as.factor)
 #'
 #' @export
 as_integer <- function(x) {
 	if (is.factor(x)) {
-		levels(x) %>% .[x] %>% as.integer
+		int <- levels(x) %>% .[x] %>% as.integer
+    if (length(na.omit(int)) == 0) as.integer(x) else int
 	} else if (is.character(x)) {
     int <- sshhr( try(as.integer(x), silent = TRUE))
     if (length(na.omit(int)) == 0) x else int